+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# 0.5.1
+Wish list:
+- When quiting the application show the accuracy report, and statistics
+- When there are no more paragraphs to type, show the accuracy report, 
+  and statistics and a message saying that the test is over.
+- Test only some fingers, meaning that it will autotype the rest of the 
+  fingers. This will be useful for practicing individual fingers.
+
+# 0.5.0:
+- Replaced `ioutil.ReadAll` with `io.ReadAll` in `main` function in `tt.go`.
+- Renamed `testFn` to `extractTypeTestFunction` in `tt.go`.
+- Changed default value for `maxLineLen` flag in `tt.go`.
+- Renamed `reflow` function to `reflowTextForScreen` in `tt.go`.
+- Updated variable names in `main` function for improved readability in `tt.go`.
+- Enhanced accuracy report precision in `showReport` function in `tt.go`.
+- Increased vertical line spacing in `start` function in `typer.go`.
+- Added typed text display with error highlights in `start` function in `typer.go`.
+- Improved text segment parameters and return types for readability in `typer.go`.
+- Refactored `start` function in `typer.go` for readability and better error management.
+- Added comment to `getParagraphs` function in `util.go`.
+- Fixed formatting in `cell` struct in `util.go`.
+- Added `.idea` to `.gitignore`.
+
 # 0.4.2:
   Added -notheme, -blockcursor and -bold.